+++ src/interrogate/cli.py
@@ -203,6 +203,18 @@ from interrogate import utils
     ),
 )
 @click.option(
+    "--style",
+    type=click.Choice(["sphinx", "google"]),
+    default="sphinx",
+    show_default=True,
+    help=(
+        "Style of docstrings to honor.\nUsing `google` will consider a class "
+        "and its `__init__` method both covered if there is either a class-"
+        "level docstring, or an `__init__` method docstring, instead of "
+        "enforcing both. Mutually exclusive with `-i`/`--ignore-init` flag."
+    ),
+)
+@click.option(
     "-o",
     "--output",
     default=None,
@@ -323,6 +335,7 @@ def main(ctx, paths, **kwargs):
     .. versionadded:: 1.5.0 ``--badge-style``
     .. versionadded:: 1.6.0 ``--ignore-overloaded-functions``
     .. verisonadded:: 1.7.0 ``--ext``
+    .. versionadded:: 1.7.0 ``--style=sphinx|google``
 
     .. versionchanged:: 1.3.1 only generate badge if results change from
         an existing badge.
@@ -340,6 +353,14 @@ def main(ctx, paths, **kwargs):
             "The `--badge-style` option must be used along with the `-g/"
             "--generate-badge option."
         )
+    if kwargs["style"] == "google" and kwargs["ignore_init_method"]:
+        raise click.BadOptionUsage(
+            option_name="style",
+            message=(
+                "The `--ignore-init-method` flag is mutually exclusive with "
+                "`--style google`."
+            ),
+        )
     if not paths:
         paths = (os.path.abspath(os.getcwd()),)
 
@@ -353,6 +374,7 @@ def main(ctx, paths, **kwargs):
 
     conf = config.InterrogateConfig(
         color=kwargs["color"],
+        docstring_style=kwargs["style"],
         fail_under=kwargs["fail_under"],
         ignore_regex=kwargs["ignore_regex"],
         ignore_magic=kwargs["ignore_magic"],