+++ tests/functional/sample/full.py
@@ -110,3 +110,18 @@ class __PrivateClass:
     """a private class"""
 
     pass
+
+
+# Coverage % for InitDocs should be the same as ClassDocs
+class InitDocs:
+    def __init__(self):
+        """A docstring for init"""
+        pass
+
+
+# Coverage % for ClassDocs should be the same as InitDocs
+class ClassDocs:
+    """A docstring for a class"""
+
+    def __init__(self):
+        pass