+++ tests/functional/test_coverage.py
@@ -44,7 +44,7 @@ def patch_term_width(monkeypatch):
                 SAMPLE_DIR,
             ],
             {},
-            (70, 36, 34, "51.4"),
+            (74, 38, 36, "51.4"),
         ),
         ([os.path.join(SAMPLE_DIR, "partial.py")], {}, (29, 10, 19, "34.5")),
         (
@@ -52,7 +52,7 @@ def patch_term_width(monkeypatch):
                 os.path.join(SAMPLE_DIR, "full.py"),
             ],
             {"ignore_nested_functions": True},
-            (24, 24, 0, "100.0"),
+            (28, 26, 2, "92.9"),
         ),
         (
             [
@@ -66,7 +66,7 @@ def patch_term_width(monkeypatch):
                 os.path.join(SAMPLE_DIR, "full.py"),
             ],
             {"ignore_overloaded_functions": True},
-            (21, 21, 0, "100.0"),
+            (25, 23, 2, "92.0"),
         ),
         (
             [
@@ -123,7 +123,10 @@ def test_coverage_errors(capsys):
 )
 def test_print_results(level, exp_fixture_file, capsys, monkeypatch):
     """Output of test results differ by verbosity."""
-    interrogate_coverage = coverage.InterrogateCoverage(paths=[SAMPLE_DIR])
+    interrogate_config = config.InterrogateConfig(docstring_style="google")
+    interrogate_coverage = coverage.InterrogateCoverage(
+        paths=[SAMPLE_DIR], conf=interrogate_config
+    )
     results = interrogate_coverage.get_coverage()
     interrogate_coverage.print_results(
         results=results, output=None, verbosity=level
@@ -152,7 +155,9 @@ def test_print_results_omit_covered(
     level, exp_fixture_file, capsys, monkeypatch
 ):
     """Output of results differ by verbosity, omitting fully covered files."""
-    interrogate_config = config.InterrogateConfig(omit_covered_files=True)
+    interrogate_config = config.InterrogateConfig(
+        omit_covered_files=True, docstring_style="google"
+    )
     interrogate_coverage = coverage.InterrogateCoverage(
         paths=[SAMPLE_DIR], conf=interrogate_config
     )
@@ -190,7 +195,9 @@ def test_print_results_omit_none(level, capsys, monkeypatch):
 
 def test_print_results_omit_all_summary(capsys, monkeypatch):
     """Output of test results for summary view, omitting all covered files."""
-    interrogate_config = config.InterrogateConfig(omit_covered_files=True)
+    interrogate_config = config.InterrogateConfig(
+        omit_covered_files=True, docstring_style="google"
+    )
     interrogate_coverage = coverage.InterrogateCoverage(
         paths=[os.path.join(SAMPLE_DIR, "full.py")], conf=interrogate_config
     )
@@ -212,7 +219,9 @@ def test_print_results_omit_all_summary(capsys, monkeypatch):
 
 def test_print_results_omit_all_detailed(capsys, monkeypatch):
     """Show no detail view when all files are omitted from skipping covered"""
-    interrogate_config = config.InterrogateConfig(omit_covered_files=True)
+    interrogate_config = config.InterrogateConfig(
+        omit_covered_files=True, docstring_style="google"
+    )
     interrogate_coverage = coverage.InterrogateCoverage(
         paths=[os.path.join(SAMPLE_DIR, "full.py")], conf=interrogate_config
     )
@@ -240,7 +249,7 @@ def test_print_results_ignore_module(
     ignore_module, level, exp_fixture_file, capsys, monkeypatch
 ):
     """Do not print module info if ignore_module is True."""
-    conf = {"ignore_module": ignore_module}
+    conf = {"ignore_module": ignore_module, "docstring_style": "google"}
     conf = config.InterrogateConfig(**conf)
 
     interrogate_coverage = coverage.InterrogateCoverage(
@@ -264,7 +273,11 @@ def test_print_results_ignore_module(
 def test_print_results_single_file(capsys, monkeypatch):
     """Results for a single file should still list the filename."""
     single_file = os.path.join(SAMPLE_DIR, "full.py")
-    interrogate_coverage = coverage.InterrogateCoverage(paths=[single_file])
+    conf = {"docstring_style": "google"}
+    conf = config.InterrogateConfig(**conf)
+    interrogate_coverage = coverage.InterrogateCoverage(
+        paths=[single_file], conf=conf
+    )
     results = interrogate_coverage.get_coverage()
     interrogate_coverage.print_results(
         results=results, output=None, verbosity=2