+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Added
 
 * `tomli` dependency for Python versions < 3.11, making use of `tomllib` in the standard library with 3.11+ (`#150 <https://github.com/econchick/interrogate/issues/150>`_).
 * Support for ``pyi`` file extensions (and leave room for other file extensions to be added, like maybe ``ipynb``).
+* Support for Google-style docstrings for class ``__init__`` methods with new ``--style [sphinx|google]`` flag (`#128 <https://github.com/econchick/interrogate/issues/128>`_).
 
 Fixed
 ^^^^^